Relocate bootmagic logic to have single entrypoint
