+++ quantum/bootmagic/bootmagic_full.h
@@ -1,3 +1,19 @@
+/* Copyright 2021 QMK
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #pragma once
 
 /* FIXME: Add special doxygen comments for defines here. */
@@ -97,6 +113,3 @@
 #ifndef BOOTMAGIC_KEY_DEFAULT_LAYER_7
 #    define BOOTMAGIC_KEY_DEFAULT_LAYER_7 KC_7
 #endif
-
-void bootmagic(void);
-bool bootmagic_scan_keycode(uint8_t keycode);
