--- tmk_core/common/magic.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void magic(void);
