+++ tmk_core/protocol.mk
@@ -54,5 +54,9 @@ ifeq ($(strip $(XT_ENABLE)), yes)
     OPT_DEFS += -DXT_ENABLE
 endif
 
+ifeq ($(strip $(USB_HID_ENABLE)), yes)
+    include $(TMK_DIR)/protocol/usb_hid.mk
+endif
+
 # Search Path
 VPATH += $(TMK_DIR)/protocol
